ajustes 06102021
ajustes 06102021

+++ application/controllers/modificaciones/cmod_insumo.php
@@ -1679,10 +1679,10 @@ class Cmod_insumo extends CI_Controller {
                       $detalle = strval(utf8_encode(trim($datos[2]))); //// descripcion form5
                       $unidad = strval(utf8_encode(trim($datos[3]))); //// Unidad
                       $cantidad = intval(trim($datos[4])); //// Cantidad
-                      $unitario = intval(trim($datos[5])); //// Costo Unitario
+                      $unitario = floatval(trim($datos[5])); //// Costo Unitario
                       
                       $p_total=($cantidad*$unitario);
-                      $total = intval(trim($datos[6])); //// Costo Total
+                      $total = floatval(trim($datos[6])); //// Costo Total
 
                       $var=7; $sum_prog=0;
                       for ($i=1; $i <=12 ; $i++) {
